Refactor application structure by replacing tvd.js with modular files. Introduced main.js as the core script, implementing a modular architecture with separate files for state management, UI handling, file operations, and chart rendering. Updated index.html to load the new main.js module, enhancing maintainability and scalability of the codebase.
