+++ index.html
@@ -189,6 +189,6 @@
     </div>
 
     <script src="data/sampleData.js"></script>
-    <script src="js/tvd.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
